Project Implementation/RAG/requirements.txt
langchain_classic==1.0.3
langchain_community==0.4.1
langchain_core==1.2.23
langchain_huggingface==1.2.1
python-dotenv==1.2.2
streamlit